+++ module-app/src/main/java/maple/expectation/application/service/EquipmentApplicationService.java
@@ -5,7 +5,7 @@ import java.util.Optional;
 import maple.expectation.core.domain.model.character.CharacterId;
 import maple.expectation.core.domain.model.equipment.CharacterEquipment;
 import maple.expectation.core.domain.model.equipment.EquipmentData;
-import maple.expectation.domain.repository.CharacterEquipmentRepository;
+import maple.expectation.infrastructure.persistence.repository.CharacterEquipmentRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
