+++ module-app/src/main/java/maple/expectation/application/worker/CharacterAsyncService.java
@@ -3,7 +3,7 @@ package maple.expectation.application.worker;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import maple.expectation.core.domain.model.character.GameCharacter;
-import maple.expectation.domain.repository.GameCharacterRepository;
+import maple.expectation.infrastructure.persistence.repository.GameCharacterRepository;
 import maple.expectation.infrastructure.executor.LogicExecutor;
 import maple.expectation.infrastructure.executor.TaskContext;
 import org.springframework.scheduling.annotation.Async;
