+++ module-app/src/main/java/maple/expectation/config/CorePortAdapterConfig.java
@@ -18,8 +18,8 @@ import maple.expectation.core.port.out.EquipmentDataPort;
 import maple.expectation.core.port.out.ItemPricePort;
 import maple.expectation.core.port.out.PotentialStatPort;
 import maple.expectation.core.port.out.ShutdownDataPersistencePort;
-import maple.expectation.domain.repository.CharacterEquipmentRepository;
-import maple.expectation.domain.repository.CubeProbabilityRepository;
+import maple.expectation.infrastructure.persistence.repository.CharacterEquipmentRepository;
+import maple.expectation.infrastructure.persistence.repository.CubeProbabilityRepository;
 import maple.expectation.domain.v2.CubeProbability;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
