+++ module-app/src/test-legacy/java/maple/expectation/monitoring/MonitoringAlertServiceTest.java
@@ -4,7 +4,7 @@ import static org.mockito.ArgumentMatchers.*;
 import static org.mockito.BDDMockito.given;
 import static org.mockito.Mockito.*;
 
-import maple.expectation.domain.repository.RedisBufferRepository;
+import maple.expectation.infrastructure.persistence.repository.RedisBufferRepository;
 import maple.expectation.infrastructure.lock.LockStrategy;
 import maple.expectation.service.v2.alert.DiscordAlertService;
 import maple.expectation.support.IntegrationTestSupport;
