+++ module-app/src/test/java/maple/expectation/application/service/character/Issues637To644E2ETest.java
@@ -21,7 +21,7 @@ import javax.crypto.SecretKey;
 import maple.expectation.core.domain.model.character.CharacterId;
 import maple.expectation.core.domain.model.character.GameCharacter;
 import maple.expectation.core.domain.model.character.UserIgn;
-import maple.expectation.domain.repository.GameCharacterRepository;
+import maple.expectation.infrastructure.persistence.repository.GameCharacterRepository;
 import maple.expectation.error.exception.CharacterNotFoundException;
 import maple.expectation.infrastructure.character.notify.CharacterCreationListener;
 import maple.expectation.infrastructure.executor.LogicExecutor;
