+++ module-app/src/test/java/maple/expectation/application/service/cube/CubeServiceTest.java
@@ -19,7 +19,7 @@ import maple.expectation.core.calculator.CubeRateCalculator;
 import maple.expectation.core.domain.stat.StatType;
 import maple.expectation.core.dto.cube.CubeCalculationInput;
 import maple.expectation.core.dto.cube.CubeComputeKey;
-import maple.expectation.domain.repository.CubeProbabilityRepository;
+import maple.expectation.infrastructure.persistence.repository.CubeProbabilityRepository;
 import maple.expectation.core.domain.model.CubeType;
 import maple.expectation.error.exception.UnsupportedCalculationEngineException;
 import maple.expectation.infrastructure.executor.LogicExecutor;
