+++ module-app/src/test/java/maple/expectation/service/v4/EquipmentExpectationServiceV4SingleflightTest.java
@@ -14,7 +14,7 @@ import maple.expectation.core.domain.model.character.GameCharacter;
 import maple.expectation.core.domain.model.character.UserIgn;
 import maple.expectation.core.domain.model.equipment.CharacterEquipment;
 import maple.expectation.core.dto.v4.EquipmentExpectationResponseV4;
-import maple.expectation.domain.repository.GameCharacterRepository;
+import maple.expectation.infrastructure.persistence.repository.GameCharacterRepository;
 import maple.expectation.parser.EquipmentStreamingParser;
 import maple.expectation.support.IntegrationTestSupport;
 import org.junit.jupiter.api.BeforeEach;
