+++ module-app/src/test/kotlin/maple/expectation/integration/repository/GameCharacterRepositoryIntegrationTest.kt
@@ -4,7 +4,7 @@ import maple.expectation.core.domain.model.PageRequest
 import maple.expectation.core.domain.model.character.CharacterId
 import maple.expectation.core.domain.model.character.GameCharacter
 import maple.expectation.core.domain.model.character.UserIgn
-import maple.expectation.domain.repository.GameCharacterRepository
+import maple.expectation.infrastructure.persistence.repository.GameCharacterRepository
 import maple.expectation.test.RepositoryIntegrationTestBase
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.DisplayName
