+++ module-app/src/test/kotlin/maple/expectation/integration/repository/MemberRepositoryIntegrationTest.kt
@@ -1,6 +1,6 @@
 package maple.expectation.integration.repository
 
-import maple.expectation.domain.repository.MemberRepository
+import maple.expectation.infrastructure.persistence.repository.MemberRepository
 import maple.expectation.domain.v2.Member
 import maple.expectation.test.RepositoryIntegrationTestBase
 import org.assertj.core.api.Assertions.assertThat
