+++ module-infra/src/main/java/maple/expectation/application/service/cube/component/SlotDistributionBuilder.java
@@ -8,7 +8,7 @@ import lombok.extern.slf4j.Slf4j;
 import maple.expectation.config.TableMassConfig;
 import maple.expectation.core.domain.model.calculator.SparsePmf;
 import maple.expectation.core.domain.stat.StatType;
-import maple.expectation.domain.repository.CubeProbabilityRepository;
+import maple.expectation.infrastructure.persistence.repository.CubeProbabilityRepository;
 import maple.expectation.domain.v2.CubeProbability;
 import maple.expectation.core.domain.model.CubeType;
 import maple.expectation.error.exception.ProbabilityInvariantException;
