+++ module-infra/src/main/java/maple/expectation/infrastructure/like/LikeToggleService.java
@@ -7,7 +7,7 @@ import maple.expectation.core.domain.model.like.LikeToggleResult;
 import maple.expectation.core.domain.model.like.LikeToggleWithCount;
 import maple.expectation.core.port.inbound.LikeTogglePort;
 import maple.expectation.core.port.out.CharacterOcidPort;
-import maple.expectation.domain.repository.CharacterLikeRepository;
+import maple.expectation.infrastructure.persistence.repository.CharacterLikeRepository;
 import maple.expectation.error.exception.CharacterNotFoundException;
 import maple.expectation.error.exception.SelfLikeNotAllowedException;
 import maple.expectation.infrastructure.executor.LogicExecutor;
