+++ module-infra/src/main/java/maple/expectation/infrastructure/like/OcidResolutionService.java
@@ -6,7 +6,7 @@ import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import maple.expectation.core.domain.model.character.GameCharacter;
 import maple.expectation.core.port.out.CharacterOcidPort;
-import maple.expectation.domain.repository.GameCharacterRepository;
+import maple.expectation.infrastructure.persistence.repository.GameCharacterRepository;
 import maple.expectation.error.exception.CharacterNotFoundException;
 import maple.expectation.infrastructure.executor.LogicExecutor;
 import maple.expectation.infrastructure.executor.TaskContext;
