+++ module-infra/src/main/kotlin/maple/expectation/infrastructure/batch/GameCharacterMicroBatchAdapter.kt
@@ -3,7 +3,7 @@ package maple.expectation.infrastructure.batch
 import io.micrometer.core.instrument.MeterRegistry
 import jakarta.annotation.PostConstruct
 import maple.expectation.core.domain.model.character.GameCharacter
-import maple.expectation.domain.repository.GameCharacterRepository
+import maple.expectation.infrastructure.persistence.repository.GameCharacterRepository
 import maple.expectation.infrastructure.config.AdaptiveMicroBatchProperties
 import maple.expectation.infrastructure.executor.LogicExecutor
 import org.springframework.cache.Cache
