+++ module-infra/src/main/kotlin/maple/expectation/infrastructure/external/impl/FallbackHandler.kt
@@ -4,7 +4,7 @@ import com.fasterxml.jackson.databind.ObjectMapper
 import java.util.concurrent.CompletableFuture
 import maple.expectation.core.domain.model.character.CharacterId
 import maple.expectation.core.domain.model.equipment.CharacterEquipment
-import maple.expectation.domain.repository.CharacterEquipmentRepository
+import maple.expectation.infrastructure.persistence.repository.CharacterEquipmentRepository
 import maple.expectation.error.exception.CharacterNotFoundException
 import maple.expectation.error.exception.EquipmentDataProcessingException
 import maple.expectation.error.exception.ExternalServiceException
