+++ module-infra/src/main/kotlin/maple/expectation/infrastructure/persistence/repository/CubeProbabilityRepositoryImpl.kt
@@ -3,7 +3,7 @@ package maple.expectation.infrastructure.persistence.repository
 import com.fasterxml.jackson.dataformat.csv.CsvMapper
 import com.fasterxml.jackson.dataformat.csv.CsvSchema
 import jakarta.annotation.PostConstruct
-import maple.expectation.domain.repository.CubeProbabilityRepository
+import maple.expectation.infrastructure.persistence.repository.CubeProbabilityRepository
 import maple.expectation.domain.v2.CubeProbability
 import maple.expectation.core.domain.model.CubeType
 import maple.expectation.error.exception.CubeDataInitializationException
