+++ module-infra/src/main/kotlin/maple/expectation/infrastructure/persistence/repository/GameCharacterRepository.kt
@@ -1,4 +1,4 @@
-package maple.expectation.domain.repository
+package maple.expectation.infrastructure.persistence.repository
 
 import maple.expectation.core.domain.model.Page
 import maple.expectation.core.domain.model.PageRequest
