+++ module-infra/src/main/kotlin/maple/expectation/infrastructure/scheduler/ExpectationCalculationScheduler.kt
@@ -4,7 +4,7 @@ import maple.expectation.core.domain.model.Page
 import maple.expectation.core.domain.model.PageRequest
 import maple.expectation.core.domain.model.character.GameCharacter
 import maple.expectation.core.port.out.QueueWriterPort
-import maple.expectation.domain.repository.GameCharacterRepository
+import maple.expectation.infrastructure.persistence.repository.GameCharacterRepository
 import maple.expectation.infrastructure.executor.LogicExecutor
 import maple.expectation.infrastructure.executor.TaskContext
 import org.slf4j.LoggerFactory
